refactor: Improve error handling and memory management across various modules
- Enhanced JSON parsing with error checks in `BitaxeFetch`, `MiningPoolStatsFetch`, and `PriceNotify` to prevent crashes on malformed input.
- Implemented defensive programming practices in `ScreenHandler` and `NostrNotify` to handle edge cases and avoid infinite loops.
- Updated `HttpHelper` to use RAII for better resource management, ensuring HTTP clients are properly released.
- Fixed potential out-of-bounds errors in LED handling and display updates.
- Improved string handling in various functions to prevent underflows and ensure proper formatting.

+++ src/lib/mining_pool/mining_pool_stats_handler.hpp
@@ -8,4 +8,4 @@
 #endif
 
 std::array<std::string, NUM_SCREENS> parseMiningPoolStatsHashRate(const std::string& hashrate, const MiningPoolInterface& pool);
-std::array<std::string, NUM_SCREENS> parseMiningPoolStatsDailyEarnings(int sats, std::string label, const MiningPoolInterface& pool);
+std::array<std::string, NUM_SCREENS> parseMiningPoolStatsDailyEarnings(int64_t sats, std::string label, const MiningPoolInterface& pool);
